requirements.txt
langchain-core==1.2.14
langchain-openai==1.1.10
langchain-community==0.4.1
langchain-classic==1.0.1
langchain-experimental==0.4.1
langchain-tavily==0.2.17
python-dotenv==1.2.1
requests
yfinance==1.2.0
streamlit==1.54.0
fastapi
uvicorn[standard]
psycopg2-binary
